front-end/src/app/views/editar-integracoes-dialog/editar-integracoes-dialog.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
import { TranslateHttpLoader } from '@ngx-translate/http-loader';
import { HttpClient } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { CustomMaterialModule } from '../../views/material/material.module';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { RouterModule } from '@angular/router';
import { MatDialog, MAT_DIALOG_DATA } from '@angular/material';
import { LoginService } from 'src/app/services/loginService/login.service';
import { of } from 'rxjs';
import { EditarIntegracoesDialogComponent } from './editar-integracoes-dialog.component';
import { EditarIntegracaoService } from 'src/app/services/editar-integracao/editar-integracao.service';

export function HttpLoaderFactory(http: HttpClient) {
  return new TranslateHttpLoader(http);
}

describe('EditarIntegracoesDialogComponent', () => {
  let component: EditarIntegracoesDialogComponent;
  let fixture: ComponentFixture<EditarIntegracoesDialogComponent>;
  let dialogSpy: MatDialog;

  beforeEach(async(() => {
    dialogSpy = jasmine.createSpyObj('MatDialog', ['open', 'closeAll']);
    TestBed.configureTestingModule({
      declarations: [ EditarIntegracoesDialogComponent ],
      imports: [
        HttpClientTestingModule,
        CustomMaterialModule,
        FormsModule,
        ReactiveFormsModule,
        BrowserAnimationsModule,
        TranslateModule.forRoot({
          loader: {
            provide: TranslateLoader,
            useFactory: HttpLoaderFactory,
            deps: [HttpClient]
          }
        }),
        RouterModule.forRoot([])
      ],
      providers: [
        { provide: MAT_DIALOG_DATA, useValue: {integracao: {codInfo: 'Teste'},
          codFesta: 'teste'}
        },
        { provide: MatDialog, useValue: dialogSpy },
        {provide: EditarIntegracaoService, useValue: {
          editarIntegracao: () => of({}),
          setFarol: () => false,
        }}
      ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(EditarIntegracoesDialogComponent);
    component = fixture.componentInstance;
    component.integracao = {codInfo: 'teste'};
    const service: LoginService = TestBed.get(LoginService);
    service.usuarioInfo = {codUsuario: '1'};
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should get f to get form controls', () => {
    expect(component.f).toBe(component.form.controls);
  });

  it('should editarForm', () => {
    spyOn(component.editarService, 'editarIntegracao')
    .and
    .callThrough();

    spyOn(component.editarService, 'setFarol')
    .and
    .callThrough();

    component.component = {
      ngOnInit: () => {
        return true;
      }
    };
    spyOn(component.component, 'ngOnInit')
    .and
    .callThrough();

    component.codFesta = 'teste';
    component.integracao = {
      codInfo: 'teste'
    };
    const integracao = {
      codInfo: 'teste',
      codFesta: 'teste',
      terIntegrado: 'teste',
      codEvent: 'teste',
      privateToken: 'teste'
    };

    component.editarIntegracao('teste', 'teste', 'teste');

    expect(component.editarService.editarIntegracao).toHaveBeenCalledWith(integracao);
    expect(component.editarService.setFarol).toHaveBeenCalledWith(false);
    expect(component.component.ngOnInit).toHaveBeenCalled();
    expect(dialogSpy.closeAll).toHaveBeenCalled();
  });
});
